pyproject.toml
[project]
name = "peasant-simulator-tavern-edition"
version = "0.1.0"
description = "A text-based medieval tavern RPG powered by Gemini AI"
authors = [
    {name = "Otis"}
]
requires-python = "^3.11"
dependencies = [
    "google-genai (==1.64.0)",
    "python-dotenv (==1.2.1)",
    "faker (==40.5.1)",
    "pydantic (>=2.0)",
    "fastapi (>=0.133.1,<1.0.0)",
    "uvicorn[standard] (>=0.41.0,<1.0.0)",
    "aiofiles (>=25.1.0,<26.0.0)",
    "python-multipart (>=0.0.22,<1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.scripts]
start = "server.app:run_server"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)"
]
